+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 0.1.1 (Next)
 
-* Your contribution here.
+* Make realtime client ping every 30s, as recommended by slack - [@samdoiron](https://github.com/samdoiron)
 
 ### 0.1.0 (7/25/2015)